requirements.txt
cdflib==0.0.0
ignite==1.1.0
imageio==2.19.3
matplotlib==3.6.2
Pillow==9.5.0
PyYAML==6.0
tqdm==4.64.1
#torch==1.13.0
#pytorch_ignite==0.4.10
#numpy==1.23.4
#torchvision==0.13.1a0